+++ src/app/components/TaskList/tasklist.module.scss
@@ -1,0 +1,8 @@
+@use '../../../assets/styles/abstracts' as *;
+
+.task-list {
+  background-color: #fff;
+  margin-top: to-rem(24);
+  border-radius: 5px;
+  box-shadow: 0 35px 50px -15px rgb(194 195 214 / 50%);
+}
