+++ src/assets/styles/abstracts/index.scss
@@ -1,2 +1,3 @@
 @forward './mixins';
 @forward './variables';
+@forward './functions';
